+++ src/sst/__init__.py
@@ -18,4 +18,6 @@
 #
 
 
-__version__ = '0.2.3dev'
+__version__ = '0.2.4dev'
+
+DEVSERVER_PORT = '8120'  # django devserver for internal acceptance tests