+++ src/sst/scripts/run.py
@@ -28,9 +28,15 @@ import time
 import traceback
 import urllib
 
+import testtools
+
+testtools.try_import('selenium')
+
+import sst
 from sst import (
     command,
     runtests,
+    tests,
 )
 
 
@@ -42,22 +48,15 @@ def main():
 
     cleanups = []
 
-    if cmd_opts.browsermob and cmd_opts.run_tests:
-        print 'Warning: can not record local traffic from django testproject'
-
-    if cmd_opts.browsermob:
-        browsermob_process = run_browsermob_server(cmd_opts.browsermob)
-
-        def browsermob_cleanup():
-            browsermob_process.kill()
-            browsermob_process.wait()
-
-        cleanups.append(('\nkilling browsermob proxy...', browsermob_cleanup))
-
     if cmd_opts.run_tests:
         cmd_opts.dir_name = 'selftests'
-        run_django()
-        cleanups.append(('\nkilling django...', kill_django))
+        if not tests.check_devserver_port_used(sst.DEVSERVER_PORT):
+            run_django(sst.DEVSERVER_PORT)
+            cleanups.append(('\nkilling django...', kill_django, sst.DEVSERVER_PORT))
+        else:
+            print 'Error: port is in use.'
+            print 'can not launch devserver for internal tests.'
+            sys.exit(1)
 
     if cmd_opts.xserver_headless:
         from sst.xvfbdisplay import Xvfb
@@ -73,8 +72,6 @@ def main():
         print '  test directory: %r' % cmd_opts.dir_name
         print '  report format: %r' % cmd_opts.report_format
         print '  browser type: %r' % cmd_opts.browser_type
-        print '  javascript disabled: %r' % cmd_opts.javascript_disabled
-        print '  browswermob proxy launcher: %r' % cmd_opts.browsermob
         print '  shared directory: %r' % cmd_opts.shared_modules
         print '  screenshots on error: %r' % cmd_opts.screenshots_on
         print '  failfast: %r' % cmd_opts.failfast
@@ -84,14 +81,13 @@ def main():
 
     try:
         command.clear_old_results()
+        factory = runtests.browser_factories.get(cmd_opts.browser_type)
         runtests.runtests(
             args,
             test_dir=cmd_opts.dir_name,
             collect_only=cmd_opts.collect_only,
             report_format=cmd_opts.report_format,
-            browser_type=cmd_opts.browser_type,
-            javascript_disabled=cmd_opts.javascript_disabled,
-            browsermob_enabled=bool(cmd_opts.browsermob),
+            browser_factory=factory(cmd_opts.javascript_disabled),
             shared_directory=cmd_opts.shared_modules,
             screenshots_on=cmd_opts.screenshots_on,
             failfast=cmd_opts.failfast,
@@ -101,19 +97,22 @@ def main():
     finally:
 
         print '--------------------------------------------------------------'
-        for desc, cmd in cleanups:
-            # Run cleanups, displaying but not propagating exceptions
+        for cleanup in cleanups:
+            # run cleanups, displaying but not propagating exceptions
+            desc = cleanup[0]
+            cmd = cleanup[1]
+            args = cleanup[2:]
+            print desc
             try:
-                print desc
-                cmd()
+                cmd(*args)
             except Exception:
                 print traceback.format_exc()
 
 
-def run_django():
+def run_django(port):
     """Start django server for running local self-tests."""
     manage_file = './src/testproject/manage.py'
-    url = 'http://localhost:8000/'
+    url = 'http://localhost:%s/' % port
 
     if not os.path.isfile(manage_file):
         print 'Error: can not find the django testproject.'
@@ -126,7 +125,7 @@ def run_django():
         print 'Error: can not find django module.'
         print 'you must have django installed to run the test project.'
         sys.exit(1)
-    subprocess.Popen([manage_file, 'runserver'],
+    proc = subprocess.Popen([manage_file, 'runserver', port],
                      stderr=open(os.devnull, 'w'),
                      stdout=open(os.devnull, 'w')
                      )
@@ -144,42 +143,14 @@ def run_django():
                 print 'Error: can not get response from %r' % url
                 raise
     print 'django found. continuing...'
+    return proc
 
 
-def kill_django():
+def kill_django(port):
     try:
-        urllib.urlopen('http://localhost:8000/kill_django')
+        urllib.urlopen('http://localhost:%s/kill_django' % port)
     except IOError:
         pass
-
-
-def run_browsermob_server(path_to_bmob_server):
-    try:
-        process = subprocess.Popen([path_to_bmob_server, ],
-                                   stderr=open(os.devnull, 'w'),
-                                   stdout=open(os.devnull, 'w')
-                                   )
-    except Exception as e:
-        print 'Error: proxy not started'
-        print e
-        sys.exit(1)
-    print '--------------------------------------------------------------'
-    print 'waiting for browsermob proxy to come up...'
-    attempts = 30
-    for count in xrange(attempts):
-        try:
-            socket_ = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            socket_.settimeout(1)
-            socket_.connect(('localhost', 8080))
-            socket_.close()
-            break
-        except socket.error:
-            time.sleep(0.2)
-            if count >= attempts - 1:  # timeout
-                print 'Error: browsermob proxy not started'
-                sys.exit(1)
-    print 'browsermob proxy found. continuing...'
-    return process
 
 
 if __name__ == '__main__':