+++ src/sst/selftests/html5.py
@@ -1,4 +1,5 @@
 from sst.actions import *
+from sst import DEVSERVER_PORT
 
 go_to('/html5')
 
@@ -6,7 +7,7 @@ assert_textfield('email')
 write_textfield('email', 'foo@bar.com', check=True)
 
 assert_textfield('url')
-write_textfield('url', 'http://localhost:8000', check=True)
+write_textfield('url', 'http://localhost:%s' % DEVSERVER_PORT, check=True)
 
 assert_textfield('search')
 write_textfield('search', 'something', check=True)