+++ src/sst/selftests/unicode.py
@@ -1,5 +1,10 @@
 from sst.actions import *
+from sst import config
 
+# currently failing in Chrome.  
+# need to investigate and file upstream chromedriver bug.
+if config.browser_type == 'chrome':
+    skip()
 
 go_to('/')