+++ src/sst/selftests/window_close.py
@@ -1,4 +1,9 @@
 from sst.actions import *
+from sst import config
+
+# PhantomJS can not do multiple windows by design
+if config.browser_type == 'phantomjs':
+    skip()
 
 go_to('/')
 click_link('popup_link')